Added tests and correct SQL

+++ src/main/java/ru/itsjava/dao/EmailJdbc.java
@@ -2,11 +2,12 @@ package ru.itsjava.dao;
 
 import ru.itsjava.domains.Email;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
 public interface EmailJdbc {
-    long countEmailByAddress(String address);
+    BigInteger countEmailByAddress(String address);
 
     Optional<Email> getEmailById(long id);
