+++ src/main/java/ru/itsjava/dao/UserJdbc.java
@@ -2,10 +2,11 @@ package ru.itsjava.dao;
 
 import ru.itsjava.domains.User;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 public interface UserJdbc {
-    long countUserByName(String fio);
+    BigInteger countUserByName(String fio);
 
     Optional<User> getUserById(long id);
