+++ src/main/java/ru/itsjava/services/UserServiceImpl.java
@@ -6,6 +6,7 @@ import org.springframework.transaction.annotation.Transactional;
 import ru.itsjava.dao.UserJdbc;
 import ru.itsjava.domains.User;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 @RequiredArgsConstructor
@@ -16,7 +17,7 @@ public class UserServiceImpl implements UserService {
     private final UserJdbc userJdbc;
 
     @Override
-    public long countUserByName(String fio) {
+    public BigInteger countUserByName(String fio) {
         return userJdbc.countUserByName(fio);
     }
